Update README.md
updtaed my Readme.md I didn't have a link to the hosted site on there.  My files are on the other computer So it was easier to make the readme changes from here rather then get the other computer out and git commit -am and push.  this is bad practice I know. bad me.